+++ src/ViewProvider/ViewProvider/WrappingViewProvider.cs
@@ -1,6 +1,10 @@
-﻿using System;
+﻿#region LICENSE NOTICE
+//! This file is subject to the terms and conditions defined in file 'LICENSE.md', which is part of this source code package.
+#endregion
+
+
+using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Windows;
 
@@ -11,12 +15,44 @@ namespace Phoenix.UI.Wpf.Architecture.VMFirst.ViewProvider
 	/// </summary>
 	public class WrappingViewProvider : List<IViewProvider>, IViewProvider
 	{
+		#region Delegates / Events
+
+		/// <inheritdoc />
+		public event EventHandler<ViewLoadedEventArgs> ViewLoaded
+		{
+			add
+			{
+				foreach (var viewProvider in this) viewProvider.ViewLoaded += value;
+			}
+			remove
+			{
+				foreach (var viewProvider in this) viewProvider.ViewLoaded -= value;
+			}
+		}
+
+		#endregion
+
+		#region Constants
+		#endregion
+
+		#region Fields
+		#endregion
+
+		#region Properties
+		#endregion
+
+		#region (De)Constructors
+
 		/// <summary>
 		/// Constructor
 		/// </summary>
 		/// <param name="viewProviders"></param>
 		public WrappingViewProvider(IEnumerable<IViewProvider> viewProviders) : base(viewProviders) { }
 
+		#endregion
+
+		#region Methods
+		
 		/// <inheritdoc />
 		public FrameworkElement GetViewInstance<TClass>(TClass viewModel) where TClass : class
 			=> this.GetViewInstance(viewModel, viewAssembly: null);
@@ -42,5 +78,7 @@ namespace Phoenix.UI.Wpf.Architecture.VMFirst.ViewProvider
 			if (System.Diagnostics.Debugger.IsAttached) System.Diagnostics.Debugger.Break();
 			throw new ViewProviderException($"Could not find a matching view for the view model '{viewModel.GetType().FullName}' in any of the available {nameof(IViewProvider)}s.");
 		}
+
+		#endregion
 	}
 }